+++ drivers/net/wireless/intel/iwlwifi/pcie/gen1_2/internal.h
@@ -611,7 +611,7 @@ struct iwl_tso_page_info {
 
 #define IWL_TSO_PAGE_DATA_SIZE	(PAGE_SIZE - sizeof(struct iwl_tso_page_info))
 #define IWL_TSO_PAGE_INFO(addr)	\
-	((struct iwl_tso_page_info *)(((unsigned long)addr & PAGE_MASK) + \
+	((struct iwl_tso_page_info *)(((uintptr_t)addr & PAGE_MASK) + \
 				      IWL_TSO_PAGE_DATA_SIZE))
 
 int iwl_pcie_tx_init(struct iwl_trans *trans);
